update int types and TimingMillis initalized

+++ cores/arduino/HardwareSerial.h
@@ -30,7 +30,7 @@ class HardwareSerial : public Stream
     void end();
     virtual int available(void) = 0;
     virtual int peek(void) = 0;
-    virtual int read(void) = 0;
+    virtual int16_t read(void) = 0;
     virtual void flush(void) = 0;
     virtual size_t write(uint8_t) = 0;
     using Print::write; // pull in write(str) and write(buf, size) from Print
