+++ cores/arduino/Stream.h
@@ -46,7 +46,7 @@ class Stream : public Print
 
   public:
     virtual int available() = 0;
-    virtual int read() = 0;
+    virtual int16_t read() = 0;
     virtual int peek() = 0;
     virtual void flush() = 0;
