+++ libraries/Wire/Wire.h
@@ -50,7 +50,7 @@ public:
 	virtual size_t write(uint8_t);
 	virtual size_t write(const uint8_t *, size_t);
 	virtual int available(void);
-	virtual int read(void);
+	virtual int16_t read(void);
 	virtual int peek(void);
 	virtual void flush(void);
 	void onReceive(void(*)(int));
